+++ src/rules/selector-class-pattern/index.js
@@ -2,7 +2,7 @@ import resolveNestedSelector from "postcss-resolve-nested-selector"
 import selectorParser from "postcss-selector-parser"
 import _ from "lodash"
 import {
-  cssRuleHasSelectorEndingWithColon,
+  isValidRule,
   report,
   ruleMessages,
   validateOptions,
@@ -34,16 +34,7 @@ export default function (pattern, options) {
       : pattern
 
     root.walkRules(rule => {
-      if (cssRuleHasSelectorEndingWithColon(rule)) { return }
-
-      // Ignore Sass intepolation possibilities
-      if (/#{.+}/.test(rule.selector)) { return }
-
-      // Ignore called Less mixins
-      if (rule.ruleWithoutBody) { return }
-
-      // Ignore non-outputting Less mixin definitions
-      if (_.endsWith(rule.selector, ")")) { return }
+      if (!isValidRule(rule)) { return }
 
       // Only bother resolving selectors that have an interpolating &
       if (shouldResolveNestedSelectors && hasInterpolatingAmpersand(rule.selector)) {